requirements.txt
psutil>=5.9.0  # System and process utilities
pdpyras>=4.5.1  # PagerDuty Python REST API Service
fastapi>=0.104.0  # For REST API
uvicorn>=0.24.0  # ASGI server for FastAPI
pyyaml>=6.0.1  # YAML configuration file handling
python-dotenv>=1.0.0  # Environment variable management
requests>=2.31.0  # HTTP requests
SQLAlchemy>=2.0.23  # Database ORM
aiohttp>=3.9.1  # Async HTTP client/server
prometheus-client>=0.19.0  # Metrics collection
pytest>=7.4.3  # Testing framework
black>=23.11.0  # Code formatting
pylint>=3.0.2  # Code linting
py-cpuinfo>=9.0.0  # CPU information 